Stub design doc
This helps with the ideation and definition of the problem at hand.
While exploring the problem, the initial idea was to port
projectile-rails. After taking a look at projectile's dev setup, I
realized that we can do this project in a didactic way, such that
newcomers to Emacs Lisp are able to contribute to the project.

I've found that the biggest bottleneck in contributing to Elisp packages is
understanding how the code is structured. Since Elisp has no namespaces
and all variables are global, the code alone cannot tell that.

The design doc emerged from both the need to formalize what we want
users to experience as well as how we want developers to feel when
working on the package.