Write projectile-project-p
We set up the initial test environment with Cask and Buttercup, as well
as a Makefile to call `cask exec` in a quicker and more convenient way.
You can call `make test` even from M-x compile.

This is one of the base functions we will need to ensure that the user
is working with a Phoenix project, since the navigation functions depend
heavily on project context. They won't work in a generic Mix project.