Write function to jump to controllers in project
With this, we can bind `projectile-phoenix-find-controller` to a key and
jump to any controller the user wishes right away.

There's a pattern to this code. The next step is generalize the
procedures to any resource we want.

As for testing, I'm finding it hard to test side-effects using
Buttercup, like opening a buffer. Ideally, we'd have some kind of return
to match, but unfortunately it does not have any matchers for buffers.
So that part of the project is uncovered for now.