Add projectile-phoenix-find-template
This allows the user to browse through templates in the project in a
faster way and being aware of nested file structures.

For this we need to implement a hash for storing the file location and
its context name. Here's an example for controllers:

"sample" ->
<project_name>/lib/<project_name>_web/controllers/sample_controller.ex
"nested/example" ->
<project_name>/lib/<project_name>_web/controllers/nested/example_controller.ex
"nested/another_nest/cogs" ->
<project_name>/lib/<project_name>_web/controllers/nested/another_nest/example_controller.ex

The user selects the desired candidate as the key of the hash and
projectile-phoenix will open the value stored in that key.

This is necessary for the template navigation because most files have
similar names and are separated by different directories. The names are
usually "show", "index", "edit", "new", "create" and "destroy". Trying
to find which of them refer to what in the old version was cumbersome.