Add projectile-phoenix-find-migration
This allows the user to find a desired migration in the Phoenix project.

We had to split the hashing logic for the context-aware choices into two
distinct functions. The difference between hash-web-choices and
hash-migration-choices is the resource type they query over.

Web resources are those under the <project_base>_web directory:

- Controllers
- Channels
- Templates
- Views

There are other types of resources, like private ones:

- Migrations
- Translation files

We won't try to get those together right now because it might be the
wrong abstraction.