[WIP] Define projectile-phoenix-mode
This is the last step in making the project usable!
With this we create the minor mode, define the keybindings it should
have and keep the prefix open for others to use.

What is left for now:

- Document how to configure projectile-phoenix-mode
- Mention that the user should define the keymap prefix as in the minor
mode conventions
- Improve the docs a bit

This should close version 0.1, and we can get some feedback and do a
test-drive of the project!