Add search for context modules
This is the most difficult filtering we've had to do so far, because the
contexts directory holds all the logic modules of the project, along
with each specific piece: workers, serializers, parsers and what have
you.

We had to grab all the .ex files in there and filter out the files that
had these specific keywords in their names, but it still feels too big.

One thing to note is that behaviours have no specific nomenclature, so
they will be mixed in there. Another side effect is that you can move
to repo modules using this function.